Removed Continent enum from class since it was only member
