+++ src/main/java/legend/game/wmap/Continent.java
@@ -1,0 +1,13 @@
+package legend.game.wmap;
+
+public enum Continent {
+  SOUTH_SERDIO,
+  NORTH_SERDIO,
+  TIBEROA,
+  ILLISA_BAY,
+  MILLE_SESEAU,
+  GLORIANO,
+  DEATH_FRONTIER,
+  ENDINESS,
+  NONE
+}
