--- src/main/java/legend/game/wmap/WmapEnums.java
@@ -1,15 +1,0 @@
-package legend.game.wmap;
-
-public class WmapEnums {
-  public enum Continent {
-    SOUTH_SERDIO,
-    NORTH_SERDIO,
-    TIBEROA,
-    ILLISA_BAY,
-    MILLE_SESEAU,
-    GLORIANO,
-    DEATH_FRONTIER,
-    ENDINESS,
-    NONE
-  }
-}
